requirements.txt
appdirs==1.4.4
apturl==0.5.2
arandr==0.1.10
attrs==20.3.0
Babel==2.6.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bitstring==3.1.9
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.2
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.3.0
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
docutils==0.14
duplicity==0.8.12.0
ebaysdk==2.1.5
ecdsa==0.17.0
entrypoints==0.3
esptool==3.1
fasteners==0.14.1
feedparser==5.2.1
freezegun==0.3.11
future==0.18.2
gevent==1.5.0
greenlet==0.4.15
html2text==2018.1.9
httplib2==0.14.0
idna==2.6
imutils==0.5.4
isodate==0.6.0
Jinja2==2.10.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libsass==0.17.0
lockfile==0.12.2
louis==3.12.0
lxml==4.3.2
macaroonbakery==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
matplotlib==3.1.2
monotonic==1.5
netifaces==0.10.4
num2words==0.5.6
numpy==1.17.4
oauthlib==3.1.0
ofxparse==0.19
olefile==0.46
paramiko==2.6.0
passlib==1.7.1
pdfminer3k==1.3.4
pexpect==4.6.0
Pillow==8.3.2
pipreqs==0.4.10
Pivy==0.6.5
ply==3.11
polib==1.1.0
powerline-status==2.7
protobuf==3.6.1
psutil==5.6.6
psycopg2==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
pydot==1.4.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.2.0
PyPDF2==1.26.0
pyRFC3339==1.1
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-ldap==3.1.0
python-stdnum==1.8
pytz==2019.1
pyusb==1.0.2
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
reedsolo==1.5.4
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
schedule==1.1.0
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.16.0
six==1.14.0
soupsieve==2.2.1
systemd-python==234
tweepy==3.10.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.3
usb-creator==0.3.7
vobject==0.9.6.1
wadllib==1.3.3
watchdog==2.0.3
Werkzeug==0.16.1
xkit==0.0.0
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
yarg==0.1.9
youtube-dl==2021.5.16
zeep==3.2.0
